Rename:Renamed game container files for consistency
