+++ client/src/components/game-container.tsx
